### 2023-02-24 #577 @diazrenata
birdsize: Estimate avian body size distributions
Submitting Author Name: Renata Diaz
Submitting Author Github Handle: @diazrenata
Repository:  https://github.com/diazrenata/birdsize
Version submitted:0.0.0.9000
Submission type: Standard
Version accepted: TBD
---

```
Package: birdsize
Title: Estimate Avian Body Size Distributions
Version: 0.0.0.9000
    person("Renata", "Diaz", , "renata.diaz@weecology.org", role = c("aut", "cre"),
Description: Generate estimated body size distributions for populations or communities of birds, given either species ID or species' mean body size. Designed to work naturally with the North American Breeding Bird Survey, or with any dataset of bird species, abundance, and/or mean size data.
    covr,
    stats

```

This package automates a workflow (seen in the wild e.g. [here](https://doi.org/10.1111/j.1466-8238.2010.00576.x) and [here](https://doi.org/10.1101/2022.11.08.515659)) of generating estimates of body size and basal metabolic rate, from the individual to ecosystem level, for birds. In my presubmission inquiry (#561, https://github.com/ropensci/software-review/issues/561), the editors determined this was best described as "field and lab reproducibility tools" because it's automating a workflow used by empirical ecologists to work with field data.

The target audience is ecologists/biodiversity scientists interested in studying the structure, function, and dynamics of bird populations and communities - specifically linking between abundances/population size and other dimensions of community function, like total biomass. Studying size-based and abundance-based properties of bird communities is key to understanding biodiversity and global change, but it is challenging for most ecologists because most survey methods do not collect size-related data. This package standardizes a computationally-intensive workaround for this challenge and makes it accessible to ecologists with relatively little computational training.

I have not encountered another package that accomplishes this.

N/A

https://github.com/ropensci/software-review/issues/561

The editor who responded to me was @annakrystalli.

As part of that conversation, the question was raised of adding the authors of some of the datasets that this package draws on as "data contributors". I agree that this is an important consideration, and I wanted to go ahead and include a little more information here so we can make sure this is done in the most appropriate way.

This package uses two sources of "external" data:

- First, the `sd_table` dataset included in the package includes (cleaned and selected) data values hand-entered from the CRC Handbook of Avian Body Masses (Dunning 2008; https://doi.org/10.1201/9781420064452). Neither Dunning, nor the authors of the studies cited in the CRC Handbook, were involved in this project. In the current iteration, I've followed the approach I would use for a paper - that is, the package and package documentation cite Dunning liberally, but I have not listed any additional authors as "data contributors"  because I generally wouldn't list folks as co-authors without their knowledge and consent. In this context, would you encourage listing Dunning as a contributor, and/or reaching out to open that conversation?

- Second, this package is designed to _interface_ with the North American Breeding Bird Survey data (https://www.sciencebase.gov/catalog/item/5d65256ae4b09b198a26c1d7, doi:10.5066/P9HE8XYJ), but I have taken care not to redistribute any actual data from the Breeding Bird Survey in the package itself. The `demo_route_raw` and `demo_route_clean` data tables in `birdsize` are synthetic datasets that mimic data from the Breeding Bird Survey. That is, they have the same column names as BBS data, and valid AOU (species identifying codes) values, but the actual data values are simulated. The `bbs-data` vignette directs users to instructions for accessing the BBS data, and demonstrates using the functions in `birdsize` on BBS-*like* data using the demo routes. Again, the package cites the Breeding Bird Survey liberally, but stops short of redistributing data so as to encourage users to access and cite the creators directly.

For both of these, again, I'm happy to explore whatever approaches to citing/crediting the original data creators seems most appropriate! I'd appreciate any thoughts or guidance in this area.

N/A

- [ ] Do you intend for this package to go on CRAN? *tbd*
