### 2023-02-24 d051bac
update README with install instructions
