### 2023-02-26 #577 @maurolepore
Thanks @diazrenata for your full submission!

Before I start the search for a handling editor can you please address these two minor yet important issues?

--

- [ ] ml01. Please add "dtc" to DESCRIPTION

In the [pre-submission]((https://github.com/ropensci/software-review/issues/561)) I see:

> One thing to note is that given the package contains external data, it would be appropriate to state the source in the DESCRIPTION file under authors using author type "dtc" which stands for "data contributor".
-- @annakrystalli

But [DESCRIPTION](https://github.com/diazrenata/birdsize/blob/main/DESCRIPTION) still shows `role = c("aut", "cre")` rather than `role = c("aut", "cre", "dtc")`.

Can you please make that change or argue against it?

--

- [ ] ml02. Please make README standalone

On the package website I see a lot of documentation and examples, well done! However the landing page is [README](https://github.com/diazrenata/birdsize/blob/main/README.md) and I see very little there. an you please add the most important bits to README to help editors, reviewers, and users understand the package quickly?

The [author's guide](https://devguide.ropensci.org/authors-guide.html) is a bit vague about whether you should duplicate in README documentation you already have elsewhere, but previous reviews and packages in the wild suggests a self-contained README is as important to a package as an abstract is to an academic paper.

![image](https://user-images.githubusercontent.com/5856545/221406249-7d80690d-016c-4176-9829-513c17df3ce5.png)
