+++ pkgdown.yml
@@ -8,5 +8,5 @@ articles:
   populations: populations.html
   scaling: scaling.html
   species: species.html
-last_built: 2023-02-24T15:45Z
+last_built: 2023-02-27T20:37Z
 
