+++ DESCRIPTION
@@ -4,6 +4,7 @@ Version: 0.0.0.9000
 Authors@R: 
     person("Renata", "Diaz", , "renata.diaz@weecology.org", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-0803-4734"))
+Date: 2023-02-27
 Description: Generate estimated body size distributions for populations or communities of birds, given either species ID or species' mean body size. Designed to work naturally with the North American Breeding Bird Survey, or with any dataset of bird species, abundance, and/or mean size data.   
 License: MIT + file LICENSE
 URL: https://github.com/diazrenata/birdsize
