+++ README.md
@@ -12,7 +12,7 @@
 
 ## Package summary
 
-`birdsize` simulates size measurements for populations or communities of birds. Because avian populuation monitoring often relies on auditory or visual methods, and therefore does not provide information about individual body *size*, it is challenging to conduct large-scale analyses of trends in biomass, energy use, community body size structure, or other, similar patterns for these species. `birdsize` solves this problem by using species' traits to generate size estimates for birds based on species ID or user-specified parameters. It is designed to interface naturally with data downloaded from the North American Breeding Bird Survey, or to work with other datasets (including synthetic data) as long as they include *population sizes* and either *species identity* (for species found in the Breeding Bird Survey) or *species mean, and optionally standard deviation, body size* (for other species, including hypothetical or simulated species).  
+`birdsize` simulates size measurements for populations or communities of birds. Because avian populuation monitoring often relies on auditory or visual methods, and therefore does not provide information about individual body *size*, it is challenging to conduct large-scale analyses of trends in biomass, energy use, community body size structure, or other, similar patterns for these species. `birdsize` solves this problem by using species' traits to generate size estimates for birds based on species ID or user-specified parameters. It is designed to interface naturally with data downloaded from the [North American Breeding Bird Survey](https://www.pwrc.usgs.gov/bbs/), or to work with other datasets (including synthetic data) as long as they include *population sizes* and either *species identity* (for species found in the Breeding Bird Survey) or *species mean, and optionally standard deviation, body size* (for other species, including hypothetical or simulated species).  
 
 The core functions in `birdsize` apply at 3 levels of organization: species, population, and community. The `community_*` functions generate population-level estimates for numerous populations (e.g. populations of different species, or populations of the same species at different points in time or different sampling locations), and calculate summary statistics with flexible grouping structures. The `population_*` functions use species-level parameters and abundances (population sizes) to simulate individual body size and basal metabolic rate measurements to make up populations of that species, and calculate population-wide summary statistics. The `species_*` functions take information about a real or hypothetical species and generates the parameters necessary to simulate body size distributions for that species.
 
