+++ review_template.md
@@ -1,0 +1,151 @@
+<!---
+Below, please enter values for (1) submitting author GitHub handle (replacing "@github_handle@); and (2) Repository URL (replacing "https://repourl"). Values for additional package authors may also be specified, replacing "@github_handle1", "@github_handle2" - delete these if not needed. DO NOT DELETE HTML SYMBOLS (everything between "<!" and ">"). Replace only "@github_handle" and "https://repourl". This comment may be deleted once it has been read and understood.
+--->
+
+Submitting Author Name: Renata Diaz
+Submitting Author Github Handle: <!--author1-->@diazrenata<!--end-author1-->
+Repository:  <!--repourl-->https://github.com/diazrenata/birdsize<!--end-repourl-->
+Version submitted:0.0.0.9000
+Submission type: <!--submission-type-->Standard<!--end-submission-type-->
+Editor: <!--editor--> TBD <!--end-editor-->
+Reviewers: <!--reviewers-list--> TBD <!--end-reviewers-list-->
+<!--due-dates-list--><!--end-due-dates-list-->
+Archive: TBD
+Version accepted: TBD
+Language: <!--language-->en<!--end-language-->
+---
+
+
+
+-   Paste the full DESCRIPTION file inside a code block below:
+
+```
+Package: birdsize
+Title: Estimate Avian Body Size Distributions
+Version: 0.0.0.9000
+Authors@R: 
+    person("Renata", "Diaz", , "renata.diaz@weecology.org", role = c("aut", "cre"),
+           comment = c(ORCID = "0000-0003-0803-4734"))
+Description: Generate estimated body size distributions for populations or communities of birds, given either species ID or species' mean body size. Designed to work naturally with the North American Breeding Bird Survey, or with any dataset of bird species, abundance, and/or mean size data.   
+License: MIT + file LICENSE
+URL: https://github.com/diazrenata/birdsize
+BugReports: https://github.com/diazrenata/birdsize/issues
+Encoding: UTF-8
+LazyData: true
+Roxygen: list(markdown = TRUE)
+RoxygenNote: 7.2.1
+Suggests: 
+    covr,
+    ggplot2,
+    knitr,
+    rmarkdown,
+    testthat (>= 3.0.0)
+Config/testthat/edition: 3
+Depends: 
+    R (>= 2.10)
+Imports: 
+    rlang,
+    dplyr,
+    magrittr,
+    purrr,
+    stats
+VignetteBuilder: knitr
+
+```
+
+
+## Scope
+
+- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below. If you are unsure, we suggest you make a pre-submission inquiry.):
+
+	- [ ] data retrieval
+	- [ ] data extraction
+	- [ ] data munging
+	- [ ] data deposition
+    - [ ] data validation and testing
+	- [ ] workflow automation
+	- [ ] version control
+	- [ ] citation management and bibliometrics
+	- [ ] scientific software wrappers
+	- [x] field and lab reproducibility tools
+	- [ ] database software bindings
+	- [ ] geospatial data
+	- [ ] text analysis
+
+- Explain how and why the package falls under these categories (briefly, 1-2 sentences):
+
+
[…]
