### 2023-03-02 b8bc2ea
- cleaned up DESCRIPTION using `desc::desc_normalize()` - remove `remotes::install_github()` instructions from README, redundant since you have `devtools::install_github()` instructions - replaced two uses of `=` for assignment with `<-` for consistency - replaced `T/F` with `TRUE/FALSE` in two places - replaced `1:nrow()` with `seq_len(nrow())` in two places
