### 2023-03-02 #577 @mstrimas
Hi @diazrenata, I'm looking forward to reviewing the package! I won't be able to get to looking at in detail until later in March, but I took a quick skim through the code and sent a small PR with a few quick fixes. Also, I thought I'd mention a few general things now ahead of my full review in April since they may take some time to address:

- I notice some very long lines in the code, especially in roxygen documentation ([for example here](https://github.com/diazrenata/birdsize/blob/main/R/species_data_functions.R#L3)), but also in the functions ([for example here](https://github.com/diazrenata/birdsize/blob/main/R/summary_functions.R#L78)). I think these would be worth shortening to the standard 80 characters to aid readability. For comments/roxygen documentation I usually use `Code > Reflow Comment` in RStudio to help with this.
- There are a lot of blank lines in the source code, sometimes multiple in a row, that make the code a little hard to read. For example [here](https://github.com/diazrenata/birdsize/blob/main/R/summary_functions.R#L70) it's hard to know what pieces of code belong together thematically. Typically chunks of code that belong together are best with no empty lines between them and you can then separate these chunks with a single blank line, often with a comment describring what each chunk does. Looks like Maëlle also mentioned this above.
- I see you use the magrittr `%>%` a lot within package functions. I don't think there's any strict rules against that, but I typically avoid it since it can make debugging trickier by producing a confusing call stack trace if there's an error. See [this comment on the rOpenSci forum](https://discuss.ropensci.org/t/current-best-practice-on-use-of-pipe-within-package-functions/3379/2?u=mstrimas). The response notes that the native pipe `|>` avoids the issue, but I personally think just dropping pipes altogether can be cleaner in package functions. This is personal preference though, so if you really prefer keeping the pipes that's all good :) Of course, using pipes in examples, vignettes, readme, etc. is perfectly fine and I think aids readability.
