### 2023-03-03 #577 @maelle
Thanks @mstrimas!

Adding a reference about code comments https://blog.r-hub.io/2023/01/26/code-comments-self-explaining-code/
