### 2023-03-21 #577 @mstrimas
I just finished up my review. This is a concise, well-documented package and was a pleasure to review. Note that the suggestions I made in an earlier comment on this issue, and in the PR I submitted, still apply and I haven't duplicated them below.

  - The README points readers to a vignette for example usage. I think it would be useful to include at least one simple example directly in the vignette to illustrate what the package does.
  - Code chunks at the start of the vignettes where packages are loaded are currently hidden. These need to be shown on the package website so users know what packages they'll need to load when working through the vignette.
  - I've noticed you refer to functions in the vignettes using, e.g., `community_generate` rather than `community_generate()`. If you add the `()`, `pkgdown` will recognize it as a function and generate a link to the documentation for that function on the website.
  - As with the package source code, there's a lot of unnecessary white space in the code chunks of the vignettes. For example, many of the code chunks have an empty line at the beginning and end. These should be removed.
  - You have a lot of `set.seed(22)` calls throughout the vignettes. I think it would be better to set the seed once at the top of each vignette rather than repeatedly setting it.
  - In some of the vignette figures, results are shown colored by species. In these figures the legend show's the AOU code. If possible showing species name would be nice.
  - At the top of the `community` vignette, you confusingly have `demo_route_raw <- demo_route_raw`. I think what you may want here is `data(demo_route_raw)`. However, I think even that has fallen out of favor and it's best to just use the object directly (See "Warning" box at https://r-pkgs.org/data.html#sec-data-data).
  - I like the `try(species_define(genus = "taylor", species = "swift"))` example :)
  - In the scaling vignette, it feels odd that you call the internal function `birdsize:::species_estimate_sd`. If you're going to use this function in a vignette, it should be exported.

  - I see no need for acknowledgement, but am happy to be listed if the author desires.

---

This is a useful, well-documented package with a clear and well defined scope. The documentation and vignettes have plenty of examples that I think most users should easily be able to follow. In addition to suggestions I've made in an earlier comment, I have a few suggestions that I think would improve the package documentation and align the source code better with best practices.

Working through the five vignettes, I did wonder if the **vignette ordering and organization** should be re-worked. For me, I found the `population` vignette to be the most useful for understanding what the package is actually doing under the hood. After working through this vignette, the `community` vignette made more sense to me. The species vignette covers an intermediate-level topic, so probably belongs next, followed by `scaling` since it's a more advanced topic that many users won't require knowledge of. Both the README and the "Getting Started" vignette contain essentially no code. This feels like missed opportunity to shown some of the most basic functionality, e.g. simulate a population of a single species and plot a histogram and/or simulate a community. Finally, there is fair amount of duplication in the vignettes, which makes me wonder if some of them could be combined together into one? All this may just be personal preference though, and other users may find the current organization better, but something to think about.

This package relies heavily on **BBS data** in all examples and package functionality. Given its prominence, I think there could more description of what the BBS is and the structure of the dataset. I initially assumed the `bbs-data` vignette would cover this, but it mostly duplicates what's already found in the `community` vignette without providing additional explanation of the BBS. I think at the very least a brief description of the route/stop structure, sampling design, and spatial/temporal coverage of the BBS is warranted. Also, I see the fields of `demo_route_raw` are described in the help for that dataset, but I think you should point users to that help file or directly include a description of the fields in the `bbs-data` vignette. I don't think you need to get into extensive detail since all of this is explained in other places, which you've referenced, but you should provide at least some explanation.

**Defining species** in function arguments could be clarified. In the arguments to `pop_generate()` and `species_define()`, species can be identified either by AOU code or scientific name. Given that both `genus` and `species` are required, it feels more intuitive to me to use a single argument (e.g. `scientific_name`). The documentation also doesn't make it clear that `species` is the species' epithet and not the scientific or common name. As it stands, it feels like you should be able to call `pop_generate(100, species = "Selasphorus calliope")` or even `pop_generate(100, species = "Calliope Hummingbird")`.

I wonder if the `_summarize()` functions are necessary since they're essentially just calling `group_by() %>% summarize()`. Personally I'd prefer to do this directly myself with `dplyr` so I know exactly what's going on and the vignettes could demonstrate exactly how users should do it. However, I can imagine that some users aren't as comfortable with `dplyr` so these convenience functions could be useful.

The internal function `ind_draw()` seems dangerous to me since it has a while loop to get rid of negative sizes with potential to run for a very long time. This is especially true because there appears to be no checks to ensure the mean size is positive. For example the following will run indefinitely:

```
pop_generate(1000, mean_size = -1000, sd_size = 0.001)
```

At the very least, please add a check to ensure `mean_size` and `sd_size` are positive and some method to ensure the while loop won't run forever, e.g. after a certain number of iterations maybe it should stop and raise an error. Even better would be re-writing this function to use a less brute force method to ensure the sizes generated aren't negative. It's not immediately clear what that would look like, since simple solutions like take the absolute value, won't preserve the desired normal distribution.

Some additional, specific comments about the code:
- I believe the standard for documenting package datasets is to document them all in `R/data.R` rather than individual R files. Not a huge issue, but I think this would help anyone interested find the source for the documentation more easily. See https://r-pkgs.org/data.html#sec-documenting-data.
- Related to the previous point, you might consider renaming or re-grouping some of the other source files. For example, the only exported function in `R/simulate_populations.R` is `pop_generate()`, so why not call that file `R/pop_generate.R` so it's more obvious what's in the file?
- `dplyr` has deprecated or superseded a lot of the "scoped" verbs, e.g. `group_by_at()`. Please replace these as appropriate to future proof your package. See https://dplyr.tidyverse.org/reference/scoped.html
- I see some inconsistency in code styling. For example, in some places you use `if()` and in other places `if ()`. I recommend picking one method of formatting your code and sticking with it. See https://style.tidyverse.org/.
- In a few places I see you're using `as.numeric(NA)` or `is.character(NA)` (e.g. https://github.com/diazrenata/birdsize/blob/main/R/species_define.R#L56). Note that there are NAs specifically for different data types, i.e. `NA_real_` and `NA_character_`. It seems cleaner to use these rather than casting `NA` to different data types.
- There are a handful of unnecessarily nested if statements (e.g. https://github.com/diazrenata/birdsize/blob/main/R/species_define.R#L54). Anywhere you have `if (A) {if (B) {...}}` I think it's cleaner to use `if (A && B) {...}`, but that may just be personal preference.
