### 2023-03-23 #577 @maelle
Thanks a lot @mstrimas for your thoughtful review!! :pray:

As a side note on source filenames, it's also nice to align them with test filenames so if you rename one, make sure you rename the other. See https://r-pkgs.org/testing-basics.html#create-a-test and [`devtools::test_active_file()`](https://devtools.r-lib.org/reference/test.html).
