### 2023-04-03 #577 @mstrimas
@qdread Thanks for expanding on my comments about the while loop! I hadn't heard of "rejection sampling" before. `truncnorm` seems quite lightweight, so maybe that's a good option. Or keep the while loop but put in some logic so it stops after some number of iterations. Also, checking the inputs so the mean isn't negative would help. I like the suggestion of having the minimum size being user defined. Anyway, I would follow whatever @qdread suggests on this since I don't really know anything about this topic.

Regarding the comment about `.data$` throwing warnings during testing, the PR I submitted should resolve this https://github.com/diazrenata/birdsize/pull/67.
