### 2023-04-04 #577 @maelle
Thanks @qdread for your thoughtful review! :smile: :bird:
You mean using tidyverse packages increases the number of _upstream_ dependencies, not downstream, correct?
