### 2023-04-04 #577 @maelle
I thought that as a salmon I was maybe confused about the flow direction. :grin:
