### 2023-05-26 65ae0fa
change to truncnorm and add a message for implausibly tiny species
