+++ review_edits.md
@@ -1,0 +1,372 @@
+---
+editor_options: 
+  markdown: 
+    wrap: 72
+---
+
+# R1
+
+Review Comments (annotations mine :))
+
+> This is a useful, well-documented package with a clear and well
+> defined scope. The documentation and vignettes have plenty of examples
+> that I think most users should easily be able to follow. In addition
+> to suggestions I've made in an earlier comment, I have a few
+> suggestions that I think would improve the package documentation and
+> align the source code better with best practices.
+
+> Working through the five vignettes, I did wonder if the vignette
+> ordering and organization should be re-worked. For me, I found the
+> population vignette to be the most useful for understanding what the
+> package is actually doing under the hood. After working through this
+> vignette, the community vignette made more sense to me. The species
+> vignette covers an intermediate-level topic, so probably belongs next,
+> followed by scaling since it's a more advanced topic that many users
+> won't require knowledge of. Both the README and the "Getting Started"
+> vignette contain essentially no code. This feels like missed
+> opportunity to shown some of the most basic functionality, e.g.
+> simulate a population of a single species and plot a histogram and/or
+> simulate a community. Finally, there is fair amount of duplication in
+> the vignettes, which makes me wonder if some of them could be combined
+> together into one? All this may just be personal preference though,
+> and other users may find the current organization better, but
+> something to think about.
+
+Initial thought: maybe put the high-level workflow demonstration in the
+README and Getting Started and then re-focus the other vignettes to
+center more intermediate and advanced *topics* rather than the "level of
+organization" approach. So:
+
+-   high-level demo in README and/or Getting Started (think about
+    redundancy here)
+-   under-the-hood in greater detail (mostly, the content from the
+    populations vignette)
+-   scaling
+-   species
+-   integration with BBS (see below, also)
+
+> This package relies heavily on BBS data in all examples and package
+> functionality. Given its prominence, I think there could more
+> description of what the BBS is and the structure of the dataset. I
+> initially assumed the bbs-data vignette would cover this, but it
+> mostly duplicates what's already found in the community vignette
+> without providing additional explanation of the BBS. I think at the
+> very least a brief description of the route/stop structure, sampling
+> design, and spatial/temporal coverage of the BBS is warranted. Also, I
+> see the fields of demo_route_raw are described in the help for that
+> dataset, but I think you should point users to that help file or
+> directly include a description of the fields in the bbs-data vignette.
+> I don't think you need to get into extensive detail since all of this
+> is explained in other places, which you've referenced, but you should
+> provide at least some explanation.
+
+Super helpful to get a gauge of how *much* context would seem helpful
+here (as I'm deep in the too-close-to-this-to-tell vortex). To-dos:
+
+-   add some more context on BBS methods and structure. Possibly as a
+    section of the intro documentation (Getting Started), or the BBS
+    vignette.
+-   adding more explanation of the fields that are inherited in
+    demo_route_raw. Here I am reflexively hesitant to duplicate the
+    contents of the BBS metadata, but it seems like an appropriate
+    context and one where doing so will make a big difference to
+    usability.
+    -   and make sure there is a direct link to the ScienceBase
+        permalink. (I intended there to be but want to confirm).
+
+> Defining species in function arguments could be clarified. In the
+> arguments to pop_generate() and species_define(), species can be
+> identified either by AOU code or scientific name. Given that both
+> genus and species are required, it feels more intuitive to me to use a
+> single argument (e.g. scientific_name). The documentation also doesn't
[…]
