+++ DESCRIPTION
@@ -12,7 +12,7 @@ BugReports: https://github.com/diazrenata/birdsize/issues
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.3
 Suggests: 
     covr,
     ggplot2,
@@ -27,5 +27,6 @@ Imports:
     dplyr,
     magrittr,
     purrr,
-    stats
+    stats,
+    truncnorm
 VignetteBuilder: knitr
