+++ NAMESPACE
@@ -25,6 +25,7 @@ importFrom(purrr,pmap_dfr)
 importFrom(rlang,.data)
 importFrom(stats,family)
 importFrom(stats,lm)
-importFrom(stats,rnorm)
+importFrom(stats,pnorm)
 importFrom(stats,sd)
 importFrom(stats,var)
+importFrom(truncnorm,rtruncnorm)
