### 2023-05-26 a0a8c92
consolidate data documentation in data.R
