+++ R/known_species.R
@@ -1,17 +1,0 @@
-#' List of species known to `birdsize`
-#'
-#' A table of the AOU (Pardieck et al 2019), genus, and species of the 443 species with species-specific data built in to `birdsize`.
-#'
-#' @format A data frame with 443 rows and 6 variables:
-#' \describe{
-#'   \item{aou}{AOU used in Paradieck et al. (2019)}
-#'   \item{genus}{genus, from Paradieck et al. (2019)}
-#'   \item{species}{species, from Paradieck et al. (2019)}
-#'   }
-#' @references \itemize{
-#'   \item{Pardieck, K.L., Ziolkowski Jr., D.J., Lutmerding, M., Aponte, V., and Hudson, M-A.R., 2019, North American Breeding Bird Survey Dataset 1966 - 2018 (ver. 2018.0): U.S. Geological Survey, Patuxent Wildlife Research Center, https://doi.org/10.5066/P9HE8XYJ.}
-#' }
-#'
-#' @keywords export
-#'
-"known_species"
