+++ R/nontarget_species.R
@@ -1,21 +1,0 @@
-#' Which AOUs correspond to nontarget species.
-#'
-#' Downloaded using the rdataretriever in June 2022.
-#' For inclusion in package vignettes, function testing.
-#' Citation needed!
-#'
-#' @format A data frame with 263 rows and 4 variables:
-#' \describe{
-#'   \item{aou}{aou}
-#'   \item{english_common_name}{English common name}
-#'   \item{genus}{genus}
-#'   \item{species}{species}
-#' }
-#' @references \itemize{
-#'   \item{Pardieck, K. L., Ziolkowski, D. J., Lutmerding, M., Aponte, V., & Hudson, M.-A. (2019). North American Breeding Bird Survey Dataset 1966—2018, version 2018.0. U.S. Geological Survey. https://doi.org/10.5066/P9HE8XYJ}
-#' }
-#'
-#' @keywords internal
-#'
-#'
-"nontarget_species"
