+++ man/demo_route_raw.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/demo_route_raw.R
+% Please edit documentation in R/data.R
 \docType{data}
 \name{demo_route_raw}
 \alias{demo_route_raw}
