+++ man/raw_masses.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/raw_masses.R
+% Please edit documentation in R/data.R
 \docType{data}
 \name{raw_masses}
 \alias{raw_masses}
