+++ NAMESPACE
@@ -9,19 +9,15 @@ export(pop_generate)
 export(pop_summarize)
 export(species_define)
 export(species_lookup)
-importFrom(dplyr,bind_rows)
 importFrom(dplyr,distinct)
 importFrom(dplyr,filter)
 importFrom(dplyr,group_by)
 importFrom(dplyr,group_by_at)
-importFrom(dplyr,left_join)
 importFrom(dplyr,mutate)
 importFrom(dplyr,n)
-importFrom(dplyr,select)
 importFrom(dplyr,summarize)
 importFrom(dplyr,ungroup)
 importFrom(magrittr,"%>%")
-importFrom(purrr,pmap_dfr)
 importFrom(rlang,.data)
 importFrom(stats,family)
 importFrom(stats,lm)
