### 2023-05-30 b319ea1
remove remaining tidy dependencies and update tests
