+++ R/summary_functions.R
@@ -12,7 +12,6 @@
 #' @examples
 #' a_population <- pop_generate(10, mean_size= 20)
 #' pop_summarize(a_population)
-#' @importFrom dplyr group_by_at summarize ungroup n distinct
 #' @importFrom stats sd
 pop_summarize <- function(population) {
 
@@ -20,21 +19,62 @@ pop_summarize <- function(population) {
 
   id_vars <- colnames(population)[ which(!(colnames(population) %in% summarize_vars))]
 
-  if(nrow(dplyr::distinct(population[,id_vars])) > 1) {
+  if(nrow(unique(population[,id_vars])) > 1) {
     warning("`population` data frame contains multiple id groups; population-wide summary may not work as expected!")
   }
 
-  population %>%
-    dplyr::group_by_at(id_vars) %>%
-    dplyr::summarize(population_abundance = dplyr::n(),
-                     population_biomass = sum(.data$individual_mass),
-                     population_metabolic_rate = sum(.data$individual_bmr),
-                     population_mean_size = mean(.data$individual_mass),
-                     population_sd_size = sd(.data$individual_mass),
-                     population_mean_bmr = mean(.data$individual_bmr),
-                     population_sd_bmr = sd(.data$individual_bmr)) %>%
-    dplyr::ungroup()
+  # population %>%
+  #   dplyr::group_by_at(id_vars) %>%
+  #   dplyr::summarize(population_abundance = dplyr::n(),
+  #                    population_biomass = sum(.data$individual_mass),
+  #                    population_metabolic_rate = sum(.data$individual_bmr),
+  #                    population_mean_size = mean(.data$individual_mass),
+  #                    population_sd_size = sd(.data$individual_mass),
+  #                    population_mean_bmr = mean(.data$individual_bmr),
+  #                    population_sd_bmr = sd(.data$individual_bmr)) %>%
+  #   dplyr::ungroup()
 
+  unique_ids <- population[ , id_vars]
+  unique_ids <- unique(unique_ids)
+
+  out_data <- unique_ids
+  out_data$population_abundance <- NA
+  out_data$population_biomass <- NA
+  out_data$population_metabolic_rate <- NA
+  out_data$population_mean_size <- NA
+  out_data$population_sd_size <- NA
+  out_data$population_mean_bmr <- NA
+  out_data$population_sd_bmr <- NA
+
+  for(i in 1:nrow(unique_ids)) {
+
+    this_id <- unique_ids[i, ]
+
+    cnames <- colnames(unique_ids)
+
+    this_pop <- population
+
+    for(j in 1:length(cnames)) {
+
+      this_col <- this_pop[ , cnames[j]]
+      this_value <- unique_ids[i, cnames[j]]
+
+      this_match <- (this_col == this_value)
+
+      this_pop <- population[ which(is.na(this_match) | this_match), ]
+    }
+
+    out_data$population_abundance[i] <- nrow(this_pop)
+    out_data$population_biomass[i] <- sum(this_pop$individual_mass)
+    out_data$population_metabolic_rate[i] <- sum(this_pop$individual_bmr)
+    out_data$population_mean_size[i] <- mean(this_pop$individual_mass)
+    out_data$population_sd_size[i] <- sd(this_pop$individual_mass)
+    out_data$population_mean_bmr[i] <- mean(this_pop$individual_bmr)
+    out_data$population_sd_bmr[i] <- sd(this_pop$individual_bmr)
+
+  }
+
+  out_data
 }
 
 #' Compute grouped summary statistics for a community
@@ -55,7 +95,6 @@ pop_summarize <- function(population) {
 #' community_data <- community_generate(bbs_route)
[…]
