+++ NAMESPACE
@@ -9,16 +9,7 @@ export(pop_generate)
 export(pop_summarize)
 export(species_define)
 export(species_lookup)
-importFrom(dplyr,distinct)
-importFrom(dplyr,filter)
-importFrom(dplyr,group_by)
-importFrom(dplyr,group_by_at)
-importFrom(dplyr,mutate)
-importFrom(dplyr,n)
-importFrom(dplyr,summarize)
-importFrom(dplyr,ungroup)
 importFrom(magrittr,"%>%")
-importFrom(rlang,.data)
 importFrom(stats,family)
 importFrom(stats,lm)
 importFrom(stats,pnorm)

[…]
