+++ tests/testthat/test-07_simulate_community.R
@@ -6,7 +6,7 @@ test_that("error catching works", {
   no_aou <- bbs_data %>%
     dplyr::select(-aou)
 
-  expect_error(community_generate(no_aou), regexp = "At least one of `aou`, `genus` and `species`, or `mean_size` is required")
+  expect_error(community_generate(no_aou), regexp = "At least one of `aou`, `scientific_name`, or `mean_size` is required")
 
   no_abundance <- bbs_data %>%
     dplyr::select(-speciestotal)
@@ -66,7 +66,7 @@ test_that("simulation from species name works", {
 
   short_bbs_data <- bbs_data %>%
     dplyr::filter(year == 2019) %>%
-    dplyr::left_join(dplyr::select(birdsize::sd_table, aou, genus, species)) %>%
+    dplyr::left_join(dplyr::select(birdsize::sd_table, aou, scientific_name)) %>%
     dplyr::select(-aou)
 
   set.seed(22)
@@ -100,8 +100,7 @@ test_that("simulation from mean size works", {
 #  expect_true(ncol(bbs_data_sims) == 23)
   expect_true(nrow(bbs_data_sims) == sum(short_bbs_data$speciestotal))
   expect_true(all(is.na(bbs_data_sims$aou)))
-  expect_true(all(is.na(bbs_data_sims$species)))
-  expect_true(all(is.na(bbs_data_sims$genus)))
+  expect_true(all(is.na(bbs_data_sims$scientific_name)))
  # expect_true(all(round(bbs_data_sims$individual_mass[1:5]) == c(5824, 7147, 127, 121, 117)))
   expect_true(all(bbs_data_sims$sd_method == "SD estimated from mean"))
 
@@ -126,8 +125,7 @@ test_that("simulation from mean and sd size works", {
  # expect_true(ncol(bbs_data_sims) == 23)
   expect_true(nrow(bbs_data_sims) == sum(short_bbs_data$speciestotal))
   expect_true(all(is.na(bbs_data_sims$aou)))
-  expect_true(all(is.na(bbs_data_sims$species)))
-  expect_true(all(is.na(bbs_data_sims$genus)))
+  expect_true(all(is.na(bbs_data_sims$scientific_name)))
   #expect_true(all(round(bbs_data_sims$individual_mass[1:5]) == c(5824, 7147, 121, 120, 119)))
   expect_true(all(bbs_data_sims$sd_method == "Mean and SD provided"))
 
@@ -148,8 +146,7 @@ test_that("simulation from toy df works", {
   #expect_true(ncol(toy_sims) == 10)
   expect_true(nrow(toy_sims) == sum(toy_df$speciestotal))
   expect_true(all(is.na(toy_sims$aou)))
-  expect_true(all(is.na(toy_sims$species)))
-  expect_true(all(is.na(toy_sims$genus)))
+  expect_true(all(is.na(toy_sims$scientific_name)))
   expect_false(any(is.na(toy_sims$sim_species_id)))
   expect_true(all(round(toy_sims$individual_mass[1:5]) == c(96, 118, 107, 102, 99)))
   expect_true(all(toy_sims$sd_method == "SD estimated from mean"))
