+++ tests/testthat/test-01_trivial.R
@@ -1,3 +1,0 @@
-test_that("trivial", {
-  expect_true(TRUE)
-})
