+++ tests/testthat/test-03_lookup_pars_fxns.R
@@ -18,18 +18,18 @@ test_that("spID lookup works", {
 test_that("genus species lookup works", {
 
   # Look up parameters for a species that does exist
-  existing_pars <- species_lookup(genus = "selasphorus", species = "calliope")
+  existing_pars <- species_lookup(scientific_name = "selasphorus calliope")
   row.names(existing_pars) <- NULL
 
-  comparison_pars <- sd_table[ sd_table$aou == 2930, ]
+  comparison_pars <- sd_table[ sd_table$aou == 4360, ]
   row.names(comparison_pars) <- NULL
 
-  expect_identical(existing_pars, dplyr::filter(sd_table, aou == 4360))
+  expect_identical(existing_pars, comparison_pars)
 
   # Case doesn't matter
-  existing_pars <- species_lookup(genus = "SELASPHORUS", species = "cAlliOpe")
+  existing_pars <- species_lookup(scientific_name = "SELASPHORUS cAlliOpe")
 
   # You should get an error if you try to look up parameters for a species that doesn't exist
-  expect_error(species_lookup(genus = "frumious", species = "bandersnatch"), regexp = "`genus` `species` combination is invalid.")
+  expect_error(species_lookup(scientific_name = "frumious bandersnatch"), regexp = "Scientific name is invalid.")
 })
 
