+++ tests/testthat/test-04_generate_individuals.R
@@ -7,8 +7,6 @@ test_that("ind_draw errors work", {
 
   expect_error(ind_draw(species_abundance = 10.5, species_sd = 10, species_mean = 100), regexp = "`species_abundance` must be a whole number")
 
-  expect_message(ind_draw(species_sd = 2, species_mean = 2, species_abundance = 10))
-
 })
 test_that("ind_draw works given toy inputs", {
   spmean <- 100

[…]
