+++ R/simulate_community.R
@@ -67,15 +67,15 @@ community_generate <- function(community_data_table, abundance_column_name = "sp
   # # For the cols to pass in, add NA columns for any of the variables that the sim fxns can use that aren't included
   na_vars <- possible_sim_vars[ which(!(possible_sim_vars %in% community_vars_mod))]
 
-  na_table <- matrix(nrow = nrow(community_data_table), ncol = length(na_vars)) %>%
-    as.data.frame()
+  na_table <- matrix(nrow = nrow(community_data_table), ncol = length(na_vars))
+  na_table <-  as.data.frame(na_table)
   colnames(na_table) <- na_vars
 
   # Split into 2 tables, one with ID cols and one for the cols to pass in.
   ids_table <- community_data_table[,id_vars]
 
-  sim_vars_table <- community_data_table[ , c(sim_vars, "rejoining_id")] %>%
-    cbind(na_table)
+  sim_vars_table <- community_data_table[ , c(sim_vars, "rejoining_id")]
+  sim_vars_table <-  cbind(sim_vars_table, na_table)
 
   # Draw populations
   # populations <- purrr::pmap_dfr(sim_vars_table,
@@ -126,8 +126,7 @@ community_generate <- function(community_data_table, abundance_column_name = "sp
 #'
 #' @examples
 #'
-#' filter_bbs_survey(demo_route_raw) %>%
-#' head()
+#' head(filter_bbs_survey(demo_route_raw))
 #'
 
 filter_bbs_survey <- function(bbs_survey_data) {
