+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export("%>%")
 export(community_generate)
 export(community_summarize)
 export(filter_bbs_survey)
@@ -9,7 +8,6 @@ export(pop_generate)
 export(pop_summarize)
 export(species_define)
 export(species_lookup)
-importFrom(magrittr,"%>%")
 importFrom(stats,family)
 importFrom(stats,lm)
 importFrom(stats,pnorm)
