+++ DESCRIPTION
@@ -25,7 +25,6 @@ Config/testthat/edition: 3
 Depends: 
     R (>= 2.10)
 Imports: 
-    magrittr,
     stats,
     truncnorm
 VignetteBuilder: knitr
