### 2023-05-31 a2960ce
use type specific NAs
