+++ tests/testthat/test-06_bbs_data.R
@@ -1,18 +1,16 @@
 test_that("filter bbs data works", {
 
 
-  bad_bbs_data <- demo_route_raw %>%
-    dplyr::select(-aou)
+  bad_bbs_data <- demo_route_raw[ , colnames(demo_route_raw) != "AOU"]
 
-  expect_error(filter_bbs_survey(bad_bbs_data), regexp = "`aou` column is required!")
+  expect_error(filter_bbs_survey(bad_bbs_data), regexp = "`AOU` column is required!")
 
   raw_bbs_data <- demo_route_raw
 
-  bad_species <- which(raw_bbs_data$aou %in% birdsize::unidentified_species$aou |
-                         raw_bbs_data$aou %in% birdsize::nontarget_species$aou)
+  bad_species <- which(raw_bbs_data$AOU %in% birdsize::unidentified_species$AOU |
+                         raw_bbs_data$AOU %in% birdsize::nontarget_species$AOU)
 
-  clean_bbs_data <- raw_bbs_data %>%
-    filter_bbs_survey()
+  clean_bbs_data <- filter_bbs_survey(raw_bbs_data)
 
   expect_true(nrow(raw_bbs_data) - nrow(clean_bbs_data) == length(bad_species))
 
