+++ data-raw/demo_route_clean.R
@@ -2,8 +2,8 @@
 
 demo_route_raw <- demo_route_raw
 
-demo_route_clean <- demo_route_raw %>%
-  filter_bbs_survey() %>%
-  dplyr::filter(year == 1994)
+demo_route_clean <- demo_route_raw
+demo_route_clean <-  filter_bbs_survey(demo_route_clean)
+demo_route_clean <- demo_route_clean[ which(demo_route_clean$year == 1994), ]
 
 usethis::use_data(demo_route_clean, overwrite = TRUE)
