+++ man/ind_draw.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/simulate_populations.R
+% Please edit documentation in R/ind_draw.R
 \name{ind_draw}
 \alias{ind_draw}
 \title{Draw individuals to make a population.}
