+++ README.md
@@ -1,46 +1,178 @@
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
 # birdsize
 
-
 <!-- badges: start -->
-[![Codecov test coverage](https://codecov.io/gh/diazrenata/birdsize/branch/main/graph/badge.svg)](https://codecov.io/gh/diazrenata/birdsize?branch=main)
 
+[![Codecov test
+coverage](https://codecov.io/gh/diazrenata/birdsize/branch/main/graph/badge.svg)](https://codecov.io/gh/diazrenata/birdsize?branch=main)
 
-[![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.org/#wip)
+[![Project Status: WIP – Initial development is in progress, but there
+has not yet been a stable, usable release suitable for the
+public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.org/#wip)
 
 [![pkgcheck](https://github.com/diazrenata/birdsize/workflows/pkgcheck/badge.svg)](https://github.com/diazrenata/birdsize/actions?query=workflow%3Apkgcheck)
 
-[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/577_status.svg)](https://github.com/ropensci/software-review/issues/577)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/577_status.svg)](https://github.com/ropensci/software-review/issues/577)
 <!-- badges: end -->
 
 ## Package summary
 
-`birdsize` simulates size measurements for populations or communities of birds. Because avian populuation monitoring often relies on auditory or visual methods, and therefore does not provide information about individual body *size*, it is challenging to conduct large-scale analyses of trends in biomass, energy use, community body size structure, or other, similar patterns for these species. `birdsize` solves this problem by using species' traits to generate size estimates for birds based on species ID or user-specified parameters. It is designed to interface naturally with data downloaded from the [North American Breeding Bird Survey](https://www.pwrc.usgs.gov/bbs/), or to work with other datasets (including synthetic data) as long as they include *population sizes* and either *species identity* (for species found in the Breeding Bird Survey) or *species mean, and optionally standard deviation, body size* (for other species, including hypothetical or simulated species).  
+`birdsize` simulates size measurements for populations or communities of
+birds. Because avian populuation monitoring often relies on auditory or
+visual methods, and therefore does not provide information about
+individual body *size*, it is challenging to conduct large-scale
+analyses of trends in biomass, energy use, community body size
+structure, or other, similar patterns for these species. `birdsize`
+solves this problem by using species’ traits to generate size estimates
+for birds based on species ID or user-specified parameters. It is
+designed to interface naturally with data downloaded from the [North
+American Breeding Bird Survey](https://www.pwrc.usgs.gov/bbs/), or to
+work with other datasets (including synthetic data) as long as they
+include *population sizes* and either *species identity* (for species
+found in the Breeding Bird Survey) or *species mean, and optionally
+standard deviation, body size* (for other species, including
+hypothetical or simulated species).
 
[…]
