+++ man/community_generate.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/simulate_community.R
+% Please edit documentation in R/community_generate.R
 \name{community_generate}
 \alias{community_generate}
 \title{Simulate individual measurements for many populations}
