+++ man/filter_bbs_survey.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/simulate_community.R
+% Please edit documentation in R/filter_bbs_survey.R
 \name{filter_bbs_survey}
 \alias{filter_bbs_survey}
 \title{Clean raw Breeding Bird Survey survey data}
