+++ man/pop_generate.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/simulate_populations.R
+% Please edit documentation in R/pop_generate.R
 \name{pop_generate}
 \alias{pop_generate}
 \title{Simulate body masses for a population}
