+++ README.Rmd
@@ -1,0 +1,127 @@
+---
+output: github_document
+---
+
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.path = "man/figures/README-",
+  out.width = "100%"
+)
+
+```
+
+# birdsize
+
+
+<!-- badges: start -->
+[![Codecov test coverage](https://codecov.io/gh/diazrenata/birdsize/branch/main/graph/badge.svg)](https://codecov.io/gh/diazrenata/birdsize?branch=main)
+
+
+[![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostatus.org/#wip)
+
+[![pkgcheck](https://github.com/diazrenata/birdsize/workflows/pkgcheck/badge.svg)](https://github.com/diazrenata/birdsize/actions?query=workflow%3Apkgcheck)
+
+[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/577_status.svg)](https://github.com/ropensci/software-review/issues/577)
+<!-- badges: end -->
+
+## Package summary
+
+`birdsize` simulates size measurements for populations or communities of birds. Because avian populuation monitoring often relies on auditory or visual methods, and therefore does not provide information about individual body *size*, it is challenging to conduct large-scale analyses of trends in biomass, energy use, community body size structure, or other, similar patterns for these species. `birdsize` solves this problem by using species' traits to generate size estimates for birds based on species ID or user-specified parameters. It is designed to interface naturally with data downloaded from the [North American Breeding Bird Survey](https://www.pwrc.usgs.gov/bbs/), or to work with other datasets (including synthetic data) as long as they include *population sizes* and either *species identity* (for species found in the Breeding Bird Survey) or *species mean, and optionally standard deviation, body size* (for other species, including hypothetical or simulated species).  
+
+The core functions in `birdsize` apply at 3 levels of organization: species, population, and community. The `population_*` functions use species-level parameters (mean and standard deviation of body mass) and population sizes to simulate individual body size and basal metabolic rate measurements to make up populations of that species, by drawing from normal distributions. The `species_*` functions - which are, for the most part, used under-the-hood by the `population_*` and `community_*` functions - take information about a real or hypothetical species and either looks up or estimates the parameters necessary to simulate body size distributions for that species. The `community_*` functions generate population-level estimates for numerous populations (e.g. populations of different species, or populations of the same species at different points in time or different sampling locations). 
+
+## Sample use cases
+
+### Installation
+
+To install the in-development version:
+
+```{r, eval = F}
+devtools::install_github('diazrenata/birdsize')
+```
+
+```{r}
+library(birdsize)
+```
+
+### Estimating body masses for a single population
+
+The core functionality of `birdsize` is to simulate individual body mass measurements for birds given either their species ID or a mean (and possibly standard deviation) of body mass, by drawing individuals from a normal distribution. At minimum, this requires a population size (abundance) and either the species' mean body size, or the species' identifier (scientific name, or [AOU code]() for species found in the BBS. If the standard deviation of body mass is not provided, `birdsize` will estimate it based on a scaling relationship between the mean and standard deviation of body mass. 
+
+The key function for this is `pop_generate`. 
+
+Here, we generate body mass estimates for a population of a hypothetical species with a mean body size of 10g:
+
+```{r}
+set.seed(13)
+
[…]

[…]
