+++ README.md
@@ -18,41 +18,67 @@ public.](https://www.repostatus.org/badges/latest/wip.svg)](https://www.repostat
 Review](https://badges.ropensci.org/577_status.svg)](https://github.com/ropensci/software-review/issues/577)
 <!-- badges: end -->
 
-## Package summary
+## Package functionality and limitations
 
-`birdsize` simulates size measurements for populations or communities of
-birds. Because avian populuation monitoring often relies on auditory or
-visual methods, and therefore does not provide information about
-individual body *size*, it is challenging to conduct large-scale
-analyses of trends in biomass, energy use, community body size
-structure, or other, similar patterns for these species. `birdsize`
-solves this problem by using species’ traits to generate size estimates
-for birds based on species ID or user-specified parameters. It is
-designed to interface naturally with data downloaded from the [North
-American Breeding Bird Survey](https://www.pwrc.usgs.gov/bbs/), or to
-work with other datasets (including synthetic data) as long as they
-include *population sizes* and either *species identity* (for species
-found in the Breeding Bird Survey) or *species mean, and optionally
-standard deviation, body size* (for other species, including
-hypothetical or simulated species).
+`birdsize` simulates body mass (in g) measurements for birds. Because
+avian monitoring methodologies often rely on auditory or visual cues,
+datasets on bird populations and communities often lack measurements of
+body mass for individuals. Without information on individuals’ body
+*size*, it is challenging to conduct large-scale analyses of trends in
+biomass, energy use, community body size structure, or other, similar
+patterns for these species. `birdsize` addresses this need by using
+species’ traits to generate size estimates for birds based on species
+identity or user-specified parameters.
 
-The core functions in `birdsize` apply at 3 levels of organization:
-species, population, and community. The `population_*` functions use
-species-level parameters (mean and standard deviation of body mass) and
-population sizes to simulate individual body size and basal metabolic
-rate measurements to make up populations of that species, by drawing
-from normal distributions. The `species_*` functions - which are, for
-the most part, used under-the-hood by the `population_*` and
-`community_*` functions - take information about a real or hypothetical
-species and either looks up or estimates the parameters necessary to
-simulate body size distributions for that species. The `community_*`
-functions generate population-level estimates for numerous populations
-(e.g. populations of different species, or populations of the same
-species at different points in time or different sampling locations).
+Specifically, `birdsize` assumes that body mass is normally distributed
+within a species, with a species-specific mean and standard deviation.
+`birdsize` uses species-level parameters (mean and standard deviation)
+to make estimates of individual-level body mass by drawing from the
+corresponding normal distributions. Because of the (very) broad nature
+of these assumptions, `birdsize` should probably not be used to generate
+fine-scale estimates of body mass. Rather, the estimates generated
+through `birdsize` are suitable for broad-scale macroecological studies,
+or for scenario exploration (via tweaking parameters), which would
+otherwise be impossible given the logistical constraints on *collecting*
+individual body mass measurements at massive scale.
 
-## Sample use cases
+`birdsize` contains built-in species-level parameters for 450 species
+(mostly those found in North America north of Mexico). It is also
+designed to take user-provided parameters to generate body size
+estimates for species not included in the built-in data, or indeed for
+hypothetical or simulated species. `birdsize` can generate estimates for
+populations (one or many individuals, all of the same species) or
+communities (multiple populations of different species, or of the same
[…]
