### 2023-08-01 bf0f913
consolidating pop and community vignettes into getting started
