+++ man/toy_size_community.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{toy_size_community}
+\alias{toy_size_community}
+\title{Toy data frame of abundances and species mean sizes (for vignettes)}
+\format{
+A data frame with 5 rows and 3 variables:
+\describe{
+\item{mean_size}{Mean mass, in g}
+\item{abundance}{Number of individuals to simulate masses for}
+\item{sim_species_id}{ID}
+}
+}
+\usage{
+toy_size_community
+}
+\description{
+This data table is a toy data frame for the vignettes. It has abundances and AOU codes for 5 species to make up a hypothetical community.
+}
+\keyword{datasets}
