+++ vignettes/birdsize.Rmd
@@ -12,11 +12,223 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
+library(birdsize)
+library(dplyr)
+library(ggplot2)
+theme_set(theme_bw())
+set.seed(22)
 ```
 
 
-`birdsize` simulates size measurements for populations or communities of birds. It is designed to interface naturally with data downloaded from the North American Breeding Bird Survey, or to work with other datasets (including synthetic data) as long as they include *population sizes* and either *species identity* (for species found in the Breeding Bird Survey) or *species mean, and optionally standard deviation, body size* (for other species, including hypothetical or simulated species).  
+## Overview
 
-The core functions in `birdsize` apply at 3 levels of organization: species, population, and community. The `community_*` functions generate population-level estimates for numerous populations (e.g. populations of different species, or populations of the same species at different points in time or different sampling locations), and calculate summary statistics with flexible grouping structures. The `population_*` functions use species-level parameters and abundances (population sizes) to simulate individual body size and basal metabolic rate measurements to make up populations of that species, and calculate population-wide summary statistics. The `species_*` functions take information about a real or hypothetical species and generates the parameters necessary to simulate body size distributions for that species.
+`birdsize` simulates mass measurements (g) for birds. Working from the assumption that individuals of a species have body masses normally distributed around a species-wide mean and standard deviation, `birdsize` simulates individual-level body mass measurements by drawing from species-specific normal distributions. There are parameters built-in for 450 bird species, or a user may supply mean and optionally standard deviation values for additional species. 
 
- For most users, the `community` vignette will provide a high-level overview of using `birdsize` to simulate and summarize community properties. The `bbs-data` vignette illustrates how to access and work with data from the Breeding Bird Survey. The `populations` and `species` vignettes provide more detail on using these functions and may be especially useful for working with birds not in the included set of species. Finally, the size estimates in `birdsize` use a scaling relationship between the mean and standard deviation of body size for some species, which is further illustrated in the `scaling` vignette.
+The core functions in `birdsize` apply at 3 levels of organization: species, population, and community:
+
+* The `species_*` functions take information about a real or hypothetical species and look up or calculate the parameters necessary to simulate body size distributions for that species.
+* The `population_*` functions use species-level parameters and abundances (population sizes) to simulate individual body size and basal metabolic rate measurements to make up populations of that species.
+* The `community_*` functions generate individual body mass estimates for multiple populations in one go (e.g. populations of different species, or populations of the same species at different points in time or different sampling locations).
+
+## Species-level parameters
+
+In order to generate body mass estimates for a species, `birdsize` needs an *abundance value* (how many individuals to generate estimates for) and at least one of:
+
+* For one of the 450 species known to `birdsize`: the scientific name (as `"Genus species"`) or the AOU code used by the American Ornithological Union (see *LINK*). See *LINK* for a list of the species with data included in `birdsize`.
[…]
