+++ man/toy_species_name_community.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{toy_species_name_community}
+\alias{toy_species_name_community}
+\title{Toy data frame of abundances and species names (for vignettes)}
+\format{
+A data frame with 5 rows and 2 variables:
+\describe{
+\item{scientific_name}{Scientific name}
+\item{abundance}{Number of individuals to simulate masses for}
+}
+}
+\usage{
+toy_species_name_community
+}
+\description{
+This data table is a toy data frame for the vignettes. It has abundances and AOU codes for 5 species to make up a hypothetical community.
+}
+\keyword{datasets}
