### 2023-11-28 #577 @diazrenata
Hi @maelle, @mstrimas, @qdread - thank you for your patience and your very thoughtful reviews! The most recent version of the package (https://github.com/diazrenata/birdsize) incorporates your feedback. I'll provide more complete responses-to-reviews in succeeding comments.
