### 2023-11-28 #577 @diazrenata
@mstrimas, in response to your main review comment: https://github.com/ropensci/software-review/issues/577#issuecomment-1478615356

> This is a useful, well-documented package with a clear and well
> defined scope. The documentation and vignettes have plenty of examples
> that I think most users should easily be able to follow. In addition
> to suggestions I've made in an earlier comment, I have a few
> suggestions that I think would improve the package documentation and
> align the source code better with best practices.

> Working through the five vignettes, I did wonder if the vignette
> ordering and organization should be re-worked. For me, I found the
> population vignette to be the most useful for understanding what the
> package is actually doing under the hood. After working through this
> vignette, the community vignette made more sense to me. The species
> vignette covers an intermediate-level topic, so probably belongs next,
> followed by scaling since it's a more advanced topic that many users
> won't require knowledge of. Both the README and the "Getting Started"
> vignette contain essentially no code. This feels like missed
> opportunity to shown some of the most basic functionality, e.g.
> simulate a population of a single species and plot a histogram and/or
> simulate a community. Finally, there is fair amount of duplication in
> the vignettes, which makes me wonder if some of them could be combined
> together into one? All this may just be personal preference though,
> and other users may find the current organization better, but
> something to think about.

Thanks for the feedback! The vignettes have been re-organized. Most of the content from the "population" and "community" vignettes have been consolidated into the "Getting Started" vignette, with smaller examples shown in the README. The "species", "scaling", and "BBS Data" vignettes have remained as separate articles, because those are more in-depth dives into specific subtopics that may be of interest to subpopulations of users.

> This package relies heavily on BBS data in all examples and package
> functionality. Given its prominence, I think there could more
> description of what the BBS is and the structure of the dataset. I
> initially assumed the bbs-data vignette would cover this, but it
> mostly duplicates what's already found in the community vignette
> without providing additional explanation of the BBS. I think at the
> very least a brief description of the route/stop structure, sampling
> design, and spatial/temporal coverage of the BBS is warranted. Also, I
> see the fields of demo_route_raw are described in the help for that
> dataset, but I think you should point users to that help file or
> directly include a description of the fields in the bbs-data vignette.
> I don't think you need to get into extensive detail since all of this
> is explained in other places, which you've referenced, but you should
> provide at least some explanation.

This is really helpful outside perspective! In this revision:

- There is additional context on the Breeding Bird Survey, including methods and its connection to this project, in the README.
- There is a direct link from the BBS Data vignette to the help page for `demo_route_raw`, which includes (brief) explanations of the column names, and a link to the main BBS data page for further information.

> Defining species in function arguments could be clarified. In the
> arguments to pop_generate() and species_define(), species can be
> identified either by AOU code or scientific name. Given that both
> genus and species are required, it feels more intuitive to me to use a
> single argument (e.g. scientific_name). The documentation also doesn't
> make it clear that species is the species' epithet and not the
> scientific or common name. As it stands, it feels like you should be
> able to call pop_generate(100, species = "Selasphorus calliope") or
> even pop_generate(100, species = "Calliope Hummingbird").

In this version, the `genus` + `species` designation has been replaced by a single argument, `scientific_name`.

In this revision, I did _not_ try to implement name matching based on the common name. My reasoning here was that the common name would be more prone to variations in spelling, etc, and so matching common names would be another layer of computational infrastructure. This does seem like a good feature to add in later iterations, or one that I could put more thought into if it seems like it would be very widely used!

> I wonder if the \_summarize() functions are necessary since they're
> essentially just calling group_by() %\>% summarize(). Personally I'd
> prefer to do this directly myself with dplyr so I know exactly what's
> going on and the vignettes could demonstrate exactly how users should
> do it. However, I can imagine that some users aren't as comfortable
> with dplyr so these convenience functions could be useful.

I agree. In this version, the `*summarize` functions are removed. I added a short demo of computing summary values using `dplyr` in the Getting Started vignette. (This also made it easier to remove tidyverse dependencies, see later comments!)

> The internal function ind_draw() seems dangerous to me since it has a
> while loop to get rid of negative sizes with potential to run for a
> very long time. This is especially true because there appears to be no
> checks to ensure the mean size is positive. For example the following
> will run indefinitely:

> pop_generate(1000, mean_size = -1000, sd_size = 0.001)

> At the very least, please add a check to ensure mean_size and sd_size
> are positive and some method to ensure the while loop won't run
> forever, e.g. after a certain number of iterations maybe it should
> stop and raise an error. Even better would be re-writing this function
> to use a less brute force method to ensure the sizes generated aren't
> negative. It's not immediately clear what that would look like, since
> simple solutions like take the absolute value, won't preserve the
> desired normal distribution.

Wow, thank you for recognizing this! This completely did not occur to me, but is of course a significant potential bug!

In this version, I've taken the suggestion to use `truncnorm::rtruncnorm` to draw from a normal distribution truncated with a lower bound of 1 (i.e. 1 gram of bird). Additionally, `pop_generate` now issues a message if the combined mean and standard deviations provided are likely to produce negative values.

> Some additional, specific comments about the code:

> ```
> I believe the standard for documenting package datasets is to document them all in R/data.R rather than individual R files. Not a huge issue, but I think this would help anyone interested find the source for the documentation more easily. See https://r-pkgs.org/data.html#sec-documenting-data.
> ```

Ah, ok! The dataset documentation is now all in R/data.R.

> ```
> Related to the previous point, you might consider renaming or re-grouping some of the other source files. For example, the only exported function in R/simulate_populations.R is pop_generate(), so why not call that file R/pop_generate.R so it's more obvious what's in the file?
> ```

I've rearranged the functions and tried to create a closer match between the functions in a file and the file name. I've kept some single-function scripts (ind_draw.R, pop_generate.R) and grouped others thematically. I hope this is easier to navigate?

> dplyr has deprecated or superseded a lot of the "scoped" verbs, e.g.
> group_by_at(). Please replace these as appropriate to future proof
> your package. See <https://dplyr.tidyverse.org/reference/scoped.html>

Yikes, and this is good to keep in mind. I've removed all `dplyr` verbs except for those in the vignette.

> I see some inconsistency in code styling. For example, in some places
> you use if() and in other places if (). I recommend picking one method
> of formatting your code and sticking with it. See
> <https://style.tidyverse.org/>.

I've reformatted the code using `styler`.

> In a few places I see you're using as.numeric(NA) or is.character(NA)
> (e.g.
> <https://github.com/diazrenata/birdsize/blob/main/R/species_define.R#L56>).
> Note that there are NAs specifically for different data types, i.e.
> NA_real\_ and NA_character\_. It seems cleaner to use these rather
> than casting NA to different data types.

I've replaced nonspecific NAs with data-type-specific NAs.

> There are a handful of unnecessarily nested if statements (e.g.
> <https://github.com/diazrenata/birdsize/blob/main/R/species_define.R#L54>).
> Anywhere you have if (A) {if (B) {...}} I think it's cleaner to use if
> (A && B) {...}, but that may just be personal preference.

I've kept these mostly as written, perhaps because it's just much easier for me to parse.
I'm happy to revisit this!
