### 2023-11-28 #577 @diazrenata
@mstrimas, regarding the Small Fixes PR:  Because of the scope of revisions, I ended up not merging that pull request and instead implementing the same changes on the revised codebase. Specifically, I tidied up the DESCRIPTION using desc::desc_normalize(), removed `remotes::install_github` from the README, switched T/F to TRUE/FALSE, and switched 1:nrow() to seq_len(nrow()). = was changed to <- via the lintr cleanup, and use of .data$ was removed with the removal of tidyverse functions.

And regarding your comments here: https://github.com/ropensci/software-review/issues/577#issuecomment-1452467927
> I notice some very long lines in the code, especially in roxygen documentation ([for example here](https://github.com/diazrenata/birdsize/blob/main/R/species_data_functions.R#L3)), but also in the functions ([for example here](https://github.com/diazrenata/birdsize/blob/main/R/summary_functions.R#L78)). I think these would be worth shortening to the standard 80 characters to aid readability. For comments/roxygen documentation I usually use Code > Reflow Comment in RStudio to help with this.

I believe I've addressed this by using Reflow Comment and Reformat Code.

>   T[here](https://github.com/diazrenata/birdsize/blob/main/R/summary_functions.R#L70) are a lot of blank lines in the source code, sometimes multiple in a row, that make the code a little hard to read. For example here it's hard to know what pieces of code belong together thematically. Typically chunks of code that belong together are best with no empty lines between them and you can then separate these chunks with a single blank line, often with a comment describring what each chunk does. Looks like Maëlle also mentioned this above.

I think this has been to some extent addressed by reformatting the code, but I confess I have not done a manual removal of the whitespace. If this is still a challenge to understanding, I can go through and manually trim it!

>  I see you use the magrittr %>% a lot within package functions. I don't think there's any strict rules against that, but I typically avoid it since it can make debugging trickier by producing a confusing call stack trace if there's an error. See [this comment on the rOpenSci forum](https://discuss.ropensci.org/t/current-best-practice-on-use-of-pipe-within-package-functions/3379/2?u=mstrimas). The response notes that the native pipe |> avoids the issue, but I personally think just dropping pipes altogether can be cleaner in package functions. This is personal preference though, so if you really prefer keeping the pipes that's all good :) Of course, using pipes in examples, vignettes, readme, etc. is perfectly fine and I think aids readability.

I've completely removed `%>%`, greatly decreased the general abundance of pipes, and replaced those that remain with the base `|>`.
