### 2023-11-28 #577 @diazrenata
@qdread, in response to your review: https://github.com/ropensci/software-review/issues/577#issuecomment-1494333794

> I think the purpose of the package is well-defined. It does a few
> well-defined and related tasks, and does them well. It also has
> demonstrated use cases, which is nice. The vignettes were
> well-organized and do a good job of providing examples of all the
> functions. I've divided my comments into three categories: code style,
> documentation, and "science stuff."

> Code style review

> The DESCRIPTION file states that the package depends on R \>=2.10 but
> the tidyverse dependencies require R \>= 3.3.

With the removal of tidyverse dependencies, I _believe_ R >= 2.10 will be sufficient.

> I agree with Matt that the documentation of the package datasets
> should be in one data.R file which is a little easier to navigate
> (though this is not too relevant to users as they won't ever see that,
> but it would be helpful for contributors)

Absolutely! The dataset documentation is now consolidated in data.R.

> This is a tough one, but I generally try to shy away from having
> tidyverse dependencies in packages. For example the lines with
> dplyr::mutate() could be rewritten from

> community \<- community %\>% dplyr::mutate( richnessSpecies =
> .data\$aou, species_designator = "aou" )
>
> to
>
> community[['richnessSpecies']] \<- community[['aou']]
> community[['species_designator']] \<- "aou"

> I understand this would be a lot of work to implement because
> tidyverse is used in most of your functions. If you do not want to
> fully remove the tidyverse dependencies, the most urgent one to
> address would be to replace dplyr::group_by_at() with
> dplyr::group_by(dplyr::across()). group_by_at() has been deprecated
> and will likely be removed from future versions of dplyr. (I'm in
> agreement with Matt's advice on that point).

>
> Also, having tidyverse dependencies greatly increases the number of downstream dependencies for the package which might not be desirable.
>

These comments (and conversations outside this) encourage me to move
away from tidyverse dependencies. In this revision, the tidyverse packages are removed, except for examples shown in the vignettes.

> I also agree with Matt that it would be good to at least rewrite the
> code not to use the magrittr pipes %\>% for the reasons he cited.

This revision has removed the magrittr pipe. The examples use the new base R pipe (`|>`).

>
> In filter_bbs_survey(), package data are loaded with unidentified_species <- unidentified_species. I am not sure that is the recommended way to internally use package data. I noticed Matt also brought this up so I would follow his advice there.
>

I've removed these calls loading internal data. This has introduced the NOTES of "no visible binding for global variable".

> In simulate_populations(), the error checking routines that cause
> failure if things like mean and standard deviation aren't provided is
> one level down in the ind_draw() function. To me it would make more
> sense if the input is checked for errors right away instead of further
> down.

This version has the error checking occurring at both levels. This makes it slightly easier to trace if a user is running `pop_generate` (new name for `simulate_populations`), and keeps the guardrails on within `ind_draw` as well.

> Line 33 of species_data_functions.R: the argument data to lm() should
> be named.

Done! (Now at line 24).

> Documentation review

>
> I thought the vignettes were well-organized and do a good job of providing examples of all the functions. One suggestion is that the vignettes could have more informative titles. Just the word community does not make it clear exactly what can be learned from going through the vignette.
>

I hope this is addressed by re-organizing the vignettes so most of the general-purpose information is in "Getting Started", and the remaining vignettes have clear, specific content (BBS data, scaling relationship, species definitions).

> I also liked the Taylor Swift example but shouldn't the species be
> Taylor and the genus Swift? :-D

Noted, and updated. :D

> A user copying and pasting code from the vignettes will not be able to
> run the code containing tidyverse functions including select() and
> pmap_df() because those packages are not explicitly loaded in the
> vignette.

The package calls have been made visible in the vignettes.

> The term "AOU" should be defined somewhere - it is known to North
> Americans but not otherwise. I understand this package is most
> intended for use with BBS so users will be familiar but it can't hurt
> to define so that other people can understand what it is (maybe not
> necessarily the user but someone looking over the code). A related
> minor point: in some of the functions where AOU is an argument, it is
> in lowercase aou but I think it would be more consistent to make it
> capitalized AOU.

I've expanded references to the AOU in the vignettes and code documentation to the "AOU numeric code" and added links to the BBS page. I've also capitalized user-facing instances of "AOU".

Interestingly, in doing this I found that the BBS AOUs appear to differ from the (more recent) codes put out by the American Ornithological Union. It looks to me like the AOU itself has switched to four-letter alpha character codes, but the BBS has kept with the (original, I believe) numeric codes. So the only lookup table for the AOU numbers to species is the BBS list, which I have linked to.

> The bar plot with total biomass by AOU in the community vignette, as
> Matt mentioned, would be more informative with species names instead
> of AOU codes. Also, because the fill color legend is redundant with
> the axis labels, it would be cleaner to just label the x axis with the
> species names and suppress the legend of fill colors.

This plot has been removed with the vignette rearrangement, but I've updated the other plot legends to use scientific names instead of AOUs.

> Comments on science stuff

> I would feel better about it if you explicitly incorporated
> uncertainty in the scaling relationships. The parameters for the
> relationship between body mass mean and standard deviation are average
> expectations, so you would end up underestimating the variation in
> body mass if you only use that mean expectation to impute missing
> standard deviations. Is there any way of incorporating uncertainty in
> those scaling parameters?

This is a tough question. Generally when _I_ work with this method,
I take the estimated mass and sd as the limit of the precision of this
approach. The uncertainty is also variable among species - for some we
have estimates and some measurements; for some species many measurements
and some just one; and from different locations and times and
investigators. For these reasons, I handle these estimates as
"best-guesses suitable for broad-scale questions, but with some
significant limitations as things get precise". I've added language to this effect in the vignettes and README.

> It would be nice to allow filter_bbs_survey() to take arguments so
> that the user could customize removing specific species groups. For
> instance it would be interesting if the user could remove only
> waterbirds and keep nocturnal birds, if they were so inclined.

This is an interesting suggestion. To implement it I would need to expand the database of masses included in the package, which would mean going back to Dunning (2008) - which invites expanding the data further than (necessarily) just the Breeding Bird Survey species. I think this becomes a wider conversation about the scope of data reproduction in the package vs. shifting that onto a user. I haven't implemented it in this revision, but if it seems like a necessary value-add it would be possible.

> The real heart of the body mass distribution simulation is in the file
> simulate_populations.R in ind_draw() at line 34:

> population \<- rnorm(n = species_abundance, mean = species_mean, sd =
> species_sd)
>
> while (any(population \< 0)) { population[which(population \< 0)] \<-
> rnorm(n = sum(population \< 0), mean = species_mean, sd = species_sd)
> }

> This looks like you are doing rejection sampling to generate samples
> from a truncated normal distribution with lower truncation bound of 0.
> I think that is fine, but it should be made clear in the documentation
> that you are doing this. I might even recommend allowing the user to
> input a lower truncation bound instead of hard-coding it at 0 (the
> default could be 0 but you could allow the user to modify this). For
> example the user might want to ensure that all masses are greater than
> 2 grams (that is roughly the lowest value I got when I generated the
> body masses for a_hundred_hummingbirds). Not too many birds weigh less
> than the Calliope Hummingbird! :-) Actually, in general I don't think
> it is clear in your documentation that samples are drawn from a normal
> distribution. Yes, it is implied by the fact that the parameters are
> mean and standard deviation but I think it would be good to be
> explicit about it. I also wanted to address Matt's point that the
> while loop in the rejection sampling may run infinitely or nearly so
> if invalid input is provided such as negative body mass. It would be
> good for you to expand the error checking code to cause failure on
> body mass means that are not positive, avoiding the potential of an
> infinite (or almost infinite loop).

> To further address the point about the rejection sampling method, Matt mentioned in his review that it would be good to use a "less brute force" method to ensure the values are >0. Actually, the method of rejection sampling that you are currently using is a well-accepted method. It is used by the R package truncnorm. For almost any realistic values of body mass mean and sd, you will get very few negative values on the initial sample, so the while loop would hardly ever even be necessary. However, if you are interested in a more direct method of sampling from a truncated normal without the while loop, I would check out [this Stackoverflow thread](https://stats.stackexchange.com/questions/56747/simulate-constrained-normal-on-lower-or-upper-bound-in-r) especially [this answer](https://stats.stackexchange.com/a/510821/54923). There are also the existing functions msm::tnorm() and truncnorm::truncnorm() but you may not want to add a dependency to your package.
>

This is a really good and important observation (see the earlier response to Matt's comment)! I went ahead and switched over to `truncnorm::rtruncnorm` to avoid the negative-values problem.

I've also added sections right at the beginning of both the README and Getting Started vignette clarifying the assumption of the normal distribution.
