### 2023-11-29 #577 @qdread
@diazrenata This is great! Thanks for thoughtfully going through all the feedback and making changes. I appreciate how you took the extra effort to remove dependencies from your package. I still would like to see uncertainty in the scaling parameters but that is more of my "wish list" than anything that is necessary to change.

@maelle if you need me to review the revised package please let me know!
