### 2023-11-30 #577 @maelle
Thanks @diazrenata!

> Response: Ah, I learned to use the numbers so that tests fail informatively in order. The tests also don't - all - correspond directly to RScripts. Is this a major barrier? If so, I'm happy to revisit!

No but there's some advantage to the correspondance if you want to switch: https://www.tidyverse.org/blog/2022/02/upkeep-testthat-3/#corresponding-files-in-r-and-teststestthat

Regarding updates to the tidyverse, I was thinking the post ["Teaching the tidyverse in 2023"](https://www.tidyverse.org/blog/2023/08/teach-tidyverse-23/) might be a way to learn of some updates, even as a developer? :thinking:
