### 2023-11-30 #577 @qdread
## Reviewer Response

Thanks for thoughtfully going through all the feedback and making changes. I appreciate how you took the extra effort to remove dependencies from your package. I still would like to see uncertainty in the scaling parameters but that is more of a "wish list" than anything that is necessary to change. It is important to think about, though, because on a log scale the mean changes as the variance changes so accounting for that variance can have a big effect.

#### Final approval (post-review)

Estimated hours spent reviewing: 5.5
