### 2023-11-30 #577 @mstrimas
## Reviewer Response

Thanks for the detailed responses to all my feedback and suggestions. The package looks significantly improved! The few things you mentioned that you didn't change are very minor points anyway.

#### Final approval (post-review)

- [ X] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package.**

Estimated hours spent reviewing: 7
