### 2023-12-08 #577 @diazrenata
Thank you all @maelle @mstrimas @qdread!

@maelle, I've just pushed changes renaming the test files and adding a line to the README about the pipe in the vignettes. (I am also deleting the `manuscript` directory and the notes I was using to keep track of review changes). I'll continue working through the checklist now.
