### 2023-12-14 #577 @diazrenata
I believe I've completed most of the transfer tasks (in this PR: https://github.com/ropensci/birdsize/pull/74)! As I wrap things up:

@mstrimas, @qdread - would it be all right with you if I acknowledged you as reviewers in the Contributors section (see the "rev" role here: https://devguide.ropensci.org/building.html#authorship)?

@maelle, @ropensci/blog-editors - I'd be very happy to contribute a blog post. Let me know what your preferred timeline might be?

Thank you all for your feedback and improvements to the package!
