### 2024-01-15 #577 @steffilazerte
Hi @diazrenata, I'm Steffi, rOpenSci Community Assistant, and I'm excited to hear that you'd like to write a blog post for us!

Sorry for the delay in reaching out to you. Are you still interested/available to write a post?

If so, as mentioned above, we have two types of blog posts: either a [short introduction to it with an example](https://ropensci.org/tech-notes/) for a technical audience or [a longer post with some narrative](https://ropensci.org/blog/) about its development or something you learned, and an example of its use for a broader readership.

To contribute a post, we'll have you open a PR to the [roweb3](https://github.com/ropensci/roweb3) repository about a week or two before our publication date. I'll review the post and give you some suggestions for content and style. You'll consider my feedback, make any changes you'd like to make, and then we merge and publish!

We have a flexible timeline right now, so when every you're ready is good for us. However, if you work better with a deadline, we could set a deadline of the last week of January for the draft, expecting to publish the first week of February (or any other date that works for you).

You can read more about the details of writing an rOpenSci blog post here: https://blogguide.ropensci.org/, and don't hesitate to ask if you have any questions. You can reach me on the Slack (Steffi LaZerte; you should have gotten an invitation by now?) or via email ([sel@steffilazerte.ca](mailto:sel@steffilazerte.ca)), and/or ping me when you open the PR to roweb3 (steffilazerte).

Thanks!
